refactor(desktop): inline the predicates that exist to be tested
